DB Updater: Fix - DB update infinity run.
Remove the use of an external "WP_Background_Process" class. it may be a buggy version that cause unexpected errors.
